requirements.txt
# MCP Server Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
neo4j==5.15.0
qdrant-client==1.7.0
redis[hiredis]==5.0.1
aiofiles==23.2.1
aiohttp==3.9.1
tenacity==8.2.3
httpx==0.25.2

# NLP and ML
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4
scipy==1.11.4

# Graph and Network Analysis
networkx==3.2.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Web Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
selenium==4.16.0
pdfplumber==0.10.3
pytesseract==0.3.10
PyMuPDF==1.23.5
Pillow==10.0.1

# YouTube Processing
yt-dlp==2023.12.30
youtube-transcript-api==0.6.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
isodate==0.6.1

# Dashboard and Visualization
streamlit==1.28.2
dash==2.16.1
vis-js==0.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Monitoring
prometheus-client==0.19.0
grafana-api==1.0.3

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Kubernetes
kubernetes==28.1.0

# Other utilities
python-dotenv==1.0.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyYAML==6.0.1
Jinja2==3.1.2
click==8.1.7
rich==13.7.0
typer==0.9.0
